removed long_body directory
