--- long_body/__init__.py
@@ -1,8 +1,0 @@
-__all__ = [
-    "test_long_request",
-    "test_long_response",
-    "body_generator",
-    "test_request_wrong_length",
-    "tester",
-    "test_response_wrong_length",
-]
