--- long_body/body_generator.py
@@ -1,8 +1,0 @@
-__author__ = "Tempesta Technologies, Inc."
-__copyright__ = "Copyright (C) 2017-2018 Tempesta Technologies, Inc."
-__license__ = "GPL2"
-
-
-def generate_body(length):
-    """Generate body of specified length"""
-    return "x" * length
